refactor: group admin endpoint adapters
