+++ crates/dcc-mcp-gateway/src/gateway/admin/application/debug_response.rs
