+++ crates/dcc-mcp-gateway/src/gateway/admin/application/general.rs
@@ -4,8 +4,8 @@ use axum::http::{HeaderMap, HeaderValue, StatusCode, header};
 use axum::response::IntoResponse;
 use serde_json::json;
 
+use super::super::state::AdminState;
 use super::debug_response::DebugListQuery;
-use super::state::AdminState;
 use dcc_mcp_gateway_admin::{ADMIN_HTML, AdminLinkBuilder};
 
 fn traffic_export_filename() -> &'static str {
