+++ crates/dcc-mcp-gateway/src/gateway/admin/application/mod.rs
@@ -8,10 +8,37 @@
 pub mod analytics;
 
 #[cfg(feature = "admin")]
+pub(super) mod agent_trace;
+
+#[cfg(feature = "admin")]
+pub mod artifacts;
+
+#[cfg(feature = "admin")]
+pub(super) mod debug_response;
+
+#[cfg(feature = "admin")]
+pub(super) mod events;
+
+#[cfg(feature = "admin")]
+pub(super) mod general;
+
+#[cfg(feature = "admin")]
 pub mod governance;
 
 #[cfg(feature = "admin")]
+pub(super) mod memory;
+
+#[cfg(feature = "admin")]
+pub mod sessions;
+
+#[cfg(feature = "admin")]
 pub(super) mod skill_health;
+
+#[cfg(feature = "admin")]
+pub(super) mod skill_paths;
+
+#[cfg(feature = "admin")]
+pub mod skill_reload;
 
 #[cfg(feature = "admin")]
 pub(super) mod traffic;
